EDAC: Add scrub control feature
Add generic EDAC scrub control in order to control the memory scrubbers
in the system. The device with scrub feature registers with EDAC device
driver, which retrieves the scrub descriptor from EDAC scrub driver and
expose the sysfs scrub control attributes for a scrub instance to userspace
in /sys/bus/edac/devices/<dev-name>/scrubX/.

The common sysfs scrub control interface abstracts the control of an
arbitrary scrubbing functionality to a common set of functions.
The sysfs scrub attribute nodes would be present only if the client driver
has implemented the corresponding attribute callback function and passed
in ops to the EDAC device driver during registration.

Co-developed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Shiju Jose <shiju.jose@huawei.com>